WIP: Try upgrading to TestNG 7.3.0

+++ presto-hive/src/test/java/io/prestosql/plugin/hive/metastore/cache/TestCachingHiveMetastore.java
@@ -106,14 +106,14 @@ public class TestCachingHiveMetastore
         assertEquals(metastore.getAllDatabases(), ImmutableList.of(TEST_DATABASE));
         assertEquals(mockClient.getAccessCount(), 1);
         assertEquals(metastore.getDatabaseNamesStats().getRequestCount(), 2);
-        assertEquals(metastore.getDatabaseNamesStats().getHitRate(), 0.5);
+        assertEquals(metastore.getDatabaseNamesStats().getHitRate(), Double.valueOf(0.5));
 
         metastore.flushCache();
 
         assertEquals(metastore.getAllDatabases(), ImmutableList.of(TEST_DATABASE));
         assertEquals(mockClient.getAccessCount(), 2);
         assertEquals(metastore.getDatabaseNamesStats().getRequestCount(), 3);
-        assertEquals(metastore.getDatabaseNamesStats().getHitRate(), 1.0 / 3);
+        assertEquals(metastore.getDatabaseNamesStats().getHitRate(), Double.valueOf(1.0 / 3));
     }
 
     @Test
@@ -125,14 +125,14 @@ public class TestCachingHiveMetastore
         assertEquals(metastore.getAllTables(TEST_DATABASE), ImmutableList.of(TEST_TABLE));
         assertEquals(mockClient.getAccessCount(), 1);
         assertEquals(metastore.getTableNamesStats().getRequestCount(), 2);
-        assertEquals(metastore.getTableNamesStats().getHitRate(), 0.5);
+        assertEquals(metastore.getTableNamesStats().getHitRate(), Double.valueOf(0.5));
 
         metastore.flushCache();
 
         assertEquals(metastore.getAllTables(TEST_DATABASE), ImmutableList.of(TEST_TABLE));
         assertEquals(mockClient.getAccessCount(), 2);
         assertEquals(metastore.getTableNamesStats().getRequestCount(), 3);
-        assertEquals(metastore.getTableNamesStats().getHitRate(), 1.0 / 3);
+        assertEquals(metastore.getTableNamesStats().getHitRate(), Double.valueOf(1.0 / 3));
     }
 
     @Test
@@ -150,14 +150,14 @@ public class TestCachingHiveMetastore
         assertNotNull(metastore.getTable(IDENTITY, TEST_DATABASE, TEST_TABLE));
         assertEquals(mockClient.getAccessCount(), 1);
         assertEquals(metastore.getTableStats().getRequestCount(), 2);
-        assertEquals(metastore.getTableStats().getHitRate(), 0.5);
+        assertEquals(metastore.getTableStats().getHitRate(), Double.valueOf(0.5));
 
         metastore.flushCache();
 
         assertNotNull(metastore.getTable(IDENTITY, TEST_DATABASE, TEST_TABLE));
         assertEquals(mockClient.getAccessCount(), 2);
         assertEquals(metastore.getTableStats().getRequestCount(), 3);
-        assertEquals(metastore.getTableStats().getHitRate(), 1.0 / 3);
+        assertEquals(metastore.getTableStats().getHitRate(), Double.valueOf(1.0 / 3));
     }
 
     @Test
@@ -179,14 +179,14 @@ public class TestCachingHiveMetastore
         assertEquals(metastore.getPartitionNames(IDENTITY, TEST_DATABASE, TEST_TABLE).get(), expectedPartitions);
         assertEquals(mockClient.getAccessCount(), 1);
         assertEquals(metastore.getPartitionNamesStats().getRequestCount(), 2);
-        assertEquals(metastore.getPartitionNamesStats().getHitRate(), 0.5);
+        assertEquals(metastore.getPartitionNamesStats().getHitRate(), Double.valueOf(0.5));
 
         metastore.flushCache();
 
         assertEquals(metastore.getPartitionNames(IDENTITY, TEST_DATABASE, TEST_TABLE).get(), expectedPartitions);
         assertEquals(mockClient.getAccessCount(), 2);
         assertEquals(metastore.getPartitionNamesStats().getRequestCount(), 3);
-        assertEquals(metastore.getPartitionNamesStats().getHitRate(), 1.0 / 3);
+        assertEquals(metastore.getPartitionNamesStats().getHitRate(), Double.valueOf(1.0 / 3));
     }
 
     @Test
@@ -207,14 +207,14 @@ public class TestCachingHiveMetastore
         assertEquals(metastore.getPartitionNamesByParts(IDENTITY, TEST_DATABASE, TEST_TABLE, parts).get(), expectedPartitions);
         assertEquals(mockClient.getAccessCount(), 1);
         assertEquals(metastore.getPartitionFilterStats().getRequestCount(), 2);
-        assertEquals(metastore.getPartitionFilterStats().getHitRate(), 0.5);
+        assertEquals(metastore.getPartitionFilterStats().getHitRate(), Double.valueOf(0.5));
 
         metastore.flushCache();
 
         assertEquals(metastore.getPartitionNamesByParts(IDENTITY, TEST_DATABASE, TEST_TABLE, parts).get(), expectedPartitions);
         assertEquals(mockClient.getAccessCount(), 2);
         assertEquals(metastore.getPartitionFilterStats().getRequestCount(), 3);
-        assertEquals(metastore.getPartitionFilterStats().getHitRate(), 1.0 / 3);
+        assertEquals(metastore.getPartitionFilterStats().getHitRate(), Double.valueOf(1.0 / 3));
     }
 
     @Test
@@ -265,7 +265,7 @@ public class TestCachingHiveMetastore
         assertEquals(mockClient.getAccessCount(), 1);
 
         assertEquals(metastore.getRolesStats().getRequestCount(), 2);
-        assertEquals(metastore.getRolesStats().getHitRate(), 0.5);
+        assertEquals(metastore.getRolesStats().getHitRate(), Double.valueOf(0.5));
 
         metastore.flushCache();
 
@@ -283,7 +283,7 @@ public class TestCachingHiveMetastore
         assertEquals(mockClient.getAccessCount(), 4);
 
         assertEquals(metastore.getRolesStats().getRequestCount(), 5);
-        assertEquals(metastore.getRolesStats().getHitRate(), 0.2);
+        assertEquals(metastore.getRolesStats().getHitRate(), Double.valueOf(0.2));
     }
 
     @Test
@@ -298,10 +298,10 @@ public class TestCachingHiveMetastore
         assertEquals(mockClient.getAccessCount(), 2);
 
         assertEquals(metastore.getTableStatisticsStats().getRequestCount(), 1);
-        assertEquals(metastore.getTableStatisticsStats().getHitRate(), 0.0);
+        assertEquals(metastore.getTableStatisticsStats().getHitRate(), Double.valueOf(0.0));
 
         assertEquals(metastore.getTableStats().getRequestCount(), 2);
-        assertEquals(metastore.getTableStats().getHitRate(), 0.5);
+        assertEquals(metastore.getTableStats().getHitRate(), Double.valueOf(0.5));
     }
 
     @Test
@@ -319,13 +319,13 @@ public class TestCachingHiveMetastore
         assertEquals(mockClient.getAccessCount(), 3);
 
         assertEquals(metastore.getPartitionStatisticsStats().getRequestCount(), 1);
-        assertEquals(metastore.getPartitionStatisticsStats().getHitRate(), 0.0);
+        assertEquals(metastore.getPartitionStatisticsStats().getHitRate(), Double.valueOf(0.0));
 
         assertEquals(metastore.getTableStats().getRequestCount(), 3);
-        assertEquals(metastore.getTableStats().getHitRate(), 2.0 / 3);
+        assertEquals(metastore.getTableStats().getHitRate(), Double.valueOf(2.0 / 3));
 
         assertEquals(metastore.getPartitionStats().getRequestCount(), 2);
-        assertEquals(metastore.getPartitionStats().getHitRate(), 0.5);
+        assertEquals(metastore.getPartitionStats().getHitRate(), Double.valueOf(0.5));
     }
 
     @Test
